session page working with requirejs
