+++ client/app/helper/ResizeSensor.js
