+++ client/app/helper/alertify-defaults.js
