+++ client/app/helper/config.js
@@ -1,0 +1,5 @@
+// require.config({
+//     paths: {
+//         analyst: 'analyst'
+//     }
+// });
