+++ client/app/main.js
@@ -1,0 +1,10 @@
+require.config({
+    'paths': {
+        'jquery': 'helper/jquery',
+    }
+});
+
+
+require(['analystView'], function(analystView) {
+    var analystView = new analystView();
+});
