+++ client/index.html
@@ -35,6 +35,9 @@
   <script type="text/javascript" src="scripts/sail_hot.js"></script>
   <script src="scripts/path.js"></script>
 
+
+  <script data-main="scripts/config" src="scripts/require.js"></script>
+
   <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css"
         integrity="sha384-BVYiiSIFeK1dGmJRAkycuHAHRg32OmUcww7on3RYdg4Va+PmSTsz/K68vbdEjh4u" crossorigin="anonymous">
   <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap-theme.min.css"
