+++ package.json
@@ -10,10 +10,11 @@
   "devDependencies": {
     "eslint": "^3.19.0",
     "eslint-config-airbnb-base": "^11.2.0",
-    "eslint-plugin-import": "^2.7.0"
+    "eslint-plugin-import": "^2.7.0",
+    "eslint-plugin-requirejs": "^3.1.1"
   },
   "scripts": {
-    "test": "./node_modules/.bin/eslint client trusted unmask shared"
+    "test": "./node_modules/.bin/eslint client"
   },
   "repository": {
     "type": "git",
